ACE-3585: Deploy Fixed Point Zero Decimal numbers as BPEL xsd:long variables
svn path=/trunk/; revision=91123

+++ tbs-bpm/bx-product/com.tibco.bx.designtime.feature/plugins/com.tibco.bx.xpdl2bpel/src/com/tibco/bx/xpdl2bpel/util/DataTypeUtil.java
@@ -66,9 +66,9 @@ public class DataTypeUtil {
                     break;
                     
                 case BasicTypeType.FLOAT:
-                    /* Sid ACE-3585 for fixed point zero decimal place numbers, create xsd:integer variables in BPEL. */
+                    /* Sid ACE-3585 for fixed point zero decimal place numbers, create xsd:long variables in BPEL. */
                     if (basicType.getScale() != null && basicType.getScale().getValue() == 0) {
-                        xsdType = "integer";
+                        xsdType = "long";
 
                     } else {
                         xsdType = "double"; //$NON-NLS-1$
